🧹 Extract WebGL shaders from rendererCore to rendererShaders to improve readability
